2.0.1: SIMPUT can now be installed in on a MAC OSX (+updated the libwcs)

+++ extlib/wcslib/Fortran/test/ttab2.f
@@ -1,7 +1,7 @@
 *=======================================================================
 *
-* WCSLIB 4.13 - an implementation of the FITS WCS standard.
-* Copyright (C) 1995-2012, Mark Calabretta
+* WCSLIB 4.25 - an implementation of the FITS WCS standard.
+* Copyright (C) 1995-2015, Mark Calabretta
 *
 * This file is part of WCSLIB.
 *
@@ -18,17 +18,11 @@
 * You should have received a copy of the GNU Lesser General Public
 * License along with WCSLIB.  If not, see http://www.gnu.org/licenses.
 *
-* Correspondence concerning WCSLIB may be directed to:
-*   Internet email: mcalabre@atnf.csiro.au
-*   Postal address: Dr. Mark Calabretta
-*                   Australia Telescope National Facility, CSIRO
-*                   PO Box 76
-*                   Epping NSW 1710
-*                   AUSTRALIA
+* Direct correspondence concerning WCSLIB to mark@calabretta.id.au
 *
-* Author: Mark Calabretta, Australia Telescope National Facility
-* http://www.atnf.csiro.au/~mcalabre/index.html
-* $Id: ttab2.f,v 4.13.1.1 2012/03/14 07:40:38 cal103 Exp cal103 $
+* Author: Mark Calabretta, Australia Telescope National Facility, CSIRO.
+* http://www.atnf.csiro.au/people/Mark.Calabretta
+* $Id: ttab2.f,v 4.25.1.2 2015/01/06 01:02:37 mcalabre Exp mcalabre $
 *=======================================================================
 
       PROGRAM TTAB2
@@ -191,8 +185,8 @@
             CALL PGSCI (1)
 
 *           Contour labelling: bottom.
-            V0 = WORLD(2,1,1)
-            V1 = WORLD(2,NP,1)
+            V0 = REAL(WORLD(2,1,1))
+            V1 = REAL(WORLD(2,NP,1))
             IF (V0.NE.V1) THEN
               IF (ABS(INT((V1-V0)/0.2)).LT.10) THEN
                 LSTEP = 20
@@ -210,8 +204,8 @@
             END IF
 
 *           Contour labelling: left.
-            V0 = WORLD(2,1,1)
-            V1 = WORLD(2,1,NP)
+            V0 = REAL(WORLD(2,1,1))
+            V1 = REAL(WORLD(2,1,NP))
             IF (V0.NE.V1) THEN
               IF (ABS(INT((V1-V0)/0.2)).LT.10) THEN
                 LSTEP = 20
@@ -229,8 +223,8 @@
             END IF
 
 *           Contour labelling: right.
-            V0 = WORLD(2,NP,1)
-            V1 = WORLD(2,NP,NP)
+            V0 = REAL(WORLD(2,NP,1))
+            V1 = REAL(WORLD(2,NP,NP))
             IF (V0.NE.V1) THEN
               IF (ABS(INT((V1-V0)/0.2)).LT.10) THEN
                 LSTEP = 20
@@ -248,8 +242,8 @@
             END IF
 
 *           Contour labelling: top.
-            V0 = WORLD(2,1,NP)
-            V1 = WORLD(2,NP,NP)
+            V0 = REAL(WORLD(2,1,NP))
+            V1 = REAL(WORLD(2,NP,NP))
             IF (V0.NE.V1) THEN
               IF (ABS(INT((V1-V0)/0.2)).LT.10) THEN
                 LSTEP = 20
@@ -269,7 +263,7 @@
 *           Draw contours for the second coordinate element.
             DO 140 I = 1, NP
               DO 130 J = 1, NP
-                ARRAY(J,I) = WORLD(2,J,I)
+                ARRAY(J,I) = REAL(WORLD(2,J,I))
  130          CONTINUE
  140        CONTINUE
